requirements.txt
pyyaml
requests
pandas
flask
ariadne 
flask-sqlalchemy 
flask-cors
pymysql